requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
boto3==1.39.14
botocore==1.39.14
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.0.0
colorama==0.4.6
cryptography==45.0.5
distro==1.9.0
fastapi==0.116.1
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
openai==1.97.1
packaging==25.0
platformdirs==4.3.8
protobuf==5.29.5
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.16.0
snowflake-snowpark-python==1.35.0
snowflake.core==1.6.0
sortedcontainers==2.4.0
starlette==0.47.2
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
wheel==0.45.1
